Translates and bugfixing
