+++ src/config/contracts.json
@@ -1,12 +1,12 @@
 {
-  "rns": "0xcb868aeabd31e2b66f74e9a55cf064abb31a4ad5",
-  "registrar": "0x5269f5bc51cdd8aa62755c97229b7eeddd8e69a6",
-  "reverseRegistrar": "0xd25c3f94a743b93ecffecbe691beea51c3c2d9d1",
-  "publicResolver": "0x4efd25e3d348f8f25a14fb7655fba6f72edfe93a",
-  "nameResolver": "0x4b1a11bf6723e60b9d2e02aa3ece34e24bde77d9",
-  "multiChainResolver": "0x99a12be4c89cbf6cfd11d1f2c029904a7b644368",
-  "rif": "0x2acc95758f8b5f583470ba265eb685a8f45fc9d5",
+  "rns": "0x830C1a4aC017Aa9734674169B6738dc80E799A51",
+  "registrar": "0x3C267DAB14cc96381456685a217A8Cdf516C8590",
+  "reverseRegistrar": "0xcE793C42fF406e7038CBaDa77f35054DbcB7208f",
+  "publicResolver": "0xc9757292abe2844fda8ddfc58c2478285b1a2a10",
+  "nameResolver": "0x9B6e84Fe09Cd29eC765b8c82e0A9a2D7dCDFc886",
+  "multiChainResolver": "0x8d2a6d6abd2dd0e7f05604b7de6bf3ddaca1722e",
+  "rif": "0xBbb5807Ac0176192E06835B378Dba2312BDdA28D",
   "revealPeriod": 172800,
-  "fifsRegistrar": "",
-  "rskOwner": ""
+  "fifsRegistrar": "0xDc7790E79422CCAd4780A6A87D2F4E2C36554914",
+  "rskOwner": "0x36CC1F7E28E89Cfe48D6B7164c0d40BB4124dC44"
 }
