+++ src/languages/es_uy.json
@@ -31,11 +31,13 @@
   "reveal_your_bid": "revele su oferta",
   "learn_about_process": "aprenda sobre el proceso",
   "hide": "ocultar",
-  "process": "Cualquier usuario puede comenzar una subasta por cualquier nombre de dominio disponible. Es una subasta pública que sigue los principios de la subasta de Vickrey. Una subasta de Vickrey es un tipo de subasta con ofertas a sobre cerrado. Los licitadores envían ofertas por escrito sin conocer las ofertas de los demás participantes de la subasta. Quien más oferta gana, pero el precio que se paga es el de la segunda oferta más alta. Hay 4 pasos para seguir:",
-  "process_step_1": "Comience una subasta por un dominio",
-  "process_step_2": "Oferte en la subasta",
-  "process_step_3": "Revele la oferta",
-  "process_step_4": "Finalice la subasta",
+  "process": "Cualquier usuario puede registrar cualquier nombre de dominio disponible. Para registrar un nombre, hay que completar 3 pasos.",
+  "process_step_1": "Solicite el registro.",
+  "process_step_1_explanation": "Su wallet se abrirá y se le pedirá que confirme la primera de dos transacciones requeridas para el registro.",
+  "process_step_2": "Espere un minuto",
+  "process_step_2_explanation": "Espere aproximadamente un minuto: el período de espera se exige para garantizar que otra persona no haya intentado registrar el mismo nombre y para protegerlo después de su solicitud.",
+  "process_step_3": "Registrar",
+  "process_step_3_explanation": "Haga clic en \"Registrar\" y su wallet volverá a abrirse. Solo después de que se confirme la segunda transacción sabrá si obtuvo el dominio.",
   "learn_more": "conozca más",
   "mycrypto_search_title": "Busque el estado en MyCrypto",
   "mycrypto_select_network": "Seleccione la red RSK en el selector del extremo superior derecho.",
@@ -180,5 +182,13 @@
   "reverse_resolution": "resolución inversa",
   "set_reverse": "establecer inversa",
   "has_reverse": "ya establecida",
-  "notifications_new_reverse_resolution": "nueva resolución inversa registrada"
+  "notifications_new_reverse_resolution": "nueva resolución inversa registrada",
+  "notifications_migrated_domain": "su dominio ha sido migrado al nuevo registrar \"primero en entrar primero en recibir servicio\"",
+  "domain_not_available": "no disponible",
+  "start_registration_for": "iniciar el registro para",
+  "rental_period": "período de alquiler en años",
+  "commit": "solicitud de registro",
+  "notifications_registrar_committed": "se ha solicitado el registro de su dominio",
+  "notifications_registrar_revealed":  "su dominio ha sido registrado",
+  "login_explanation": "ahora puede iniciar sesión con su dominio recién registrado"
 }
