+++ src/languages/ja.json
@@ -31,11 +31,13 @@
   "reveal_your_bid": "自身の入札を公開する",
   "learn_about_process": "プロセスについて学ぶ",
   "hide": "非表示にする",
-  "process": "すべてのユーザーが利用可能なドメイン名でオークションを開始できます。そして、Vickrey オークション原則が尊重される公開オークションです。Vickrey オークションは一種の非公開オークションです。入札者はオークションにて他人の入札のことは認識せずに書面形式の入札を行います。最高値を付けた入札者が選ばれますが、支払う価格は2番目に高い値とします。手順は４つです：",
-  "process_step_1": "ドメインのオークションを開始します",
-  "process_step_2": "オークションに入札します",
-  "process_step_3": "入札を開示します",
-  "process_step_4": "オークションを確定します",
+  "process": "ユーザーは使用可能な任意のドメイン名を登録できます。1つの名前を登録するには3つのステップを完了する必要があります",
+  "process_step_1": "登録リクエスト",
+  "process_step_1_explanation": "ご使用中のウォレットが開き、登録に必要な2つのうちの最初のトランザクションを確認することを求められます。",
+  "process_step_2": "１分間待ちます",
+  "process_step_2_explanation": "おおよそ1分間ほど待ちます：\r他の人が同じ名前を登録しようとしていないかを確認し、お客様のリクエストの後でお客様を保護するためにその待ち時間が必要です。",
+  "process_step_3": "登録します",
+  "process_step_3_explanation": "「登録する」をクリックするとご使用中のウォレットが再び開きます。2番目のトランザクションを確認した後にならないと、ドメインは取得されているのかわかりません。",
   "learn_more": "詳細を確認する",
   "mycrypto_search_title": "MyCrypto のステータスを検索する",
   "mycrypto_select_network": "右上のセレクタにてRSK を選択する。",
@@ -180,5 +182,13 @@
   "reverse_resolution": "逆引き",
   "set_reverse": "逆向きに設定します",
   "has_reverse": "すでに設定されている",
-  "notifications_new_reverse_resolution": "新しい逆引きが登録されました"
+  "notifications_new_reverse_resolution": "新しい逆引きが登録されました",
+  "notifications_migrated_domain": "お客様のドメインは新しいfifsレジストラーに移行されました",
+  "domain_not_available": "使用できません",
+  "start_registration_for": "登録を開始します",
+  "rental_period": "何年ものレンタル期間",
+  "commit": "登録のリクエスト",
+  "notifications_registrar_committed": "お客様のドメインを登録する必要があります",
+  "notifications_registrar_revealed":  "お客様のドメインが登録されました",
+  "login_explanation": "お客様は登録したばかりのドメインに今すぐログインできます"
 }
