+++ src/languages/ko.json
@@ -31,11 +31,13 @@
   "reveal_your_bid": "본인의 입찰 가격 공개하기",
   "learn_about_process": "과정 알아보기",
   "hide": "숨기기",
-  "process": "모든 사용자는 사용 가능한 도메인 이름에 대해 경매를 시작할 수 있습니다. 이는 Vickrey 경매 원칙을 준수하는 공공 경매입니다. Vickrey 경매는 일종의 비공개 경매입니다. 입찰자는 다른 경매 참여자의 입찰을 알지 못한 상태에서 서면 입찰서를 제출합니다. 가장 높은 가격을 제시한 입찰자가 낙찰되지만 지불하는 가격은 두 번째로 높은 입찰자의 가격입니다. 다음과 같은 4단계로 진행됩니다.",
-  "process_step_1": "도메인에 대한 경매 시작하기",
-  "process_step_2": "경매에 입찰하기",
-  "process_step_3": "입찰서 개봉하기",
-  "process_step_4": "경매 종료",
+  "process": "모든 사용자는 사용 가능한 도메인 이름을 등록할 수 있습니다. 이름을 등록하려면 3 단계를 완료해야 합니다.",
+  "process_step_1": "등록 요청하기",
+  "process_step_1_explanation": "월렛(지갑)이 열리고 등록에 필요한 두 거래 중 첫 번째 거래를 확인하라는 메시지가 표시됩니다.",
+  "process_step_2": "1분만 기다려 주십시오",
+  "process_step_2_explanation": "약 1분 정도 기다리기: 이 대기 시간은 다른 사람이 동일한 이름을 등록하고자 했었는지를 확인하고, 등록 후 귀하를 보호하기 위해 요구됩니다.",
+  "process_step_3": "등록하기",
+  "process_step_3_explanation": "“등록하기”를 클릭하면 여러분의 월렛(지갑)이 다시 열립니다. 2번째 거래가 확인된 후에만 도메인을 획득했는지를 알 수 있습니다.",
   "learn_more": "더 알아보기",
   "mycrypto_search_title": "MyCrypto에서 상태 검색",
   "mycrypto_select_network": "오른쪽 상단의 선택자에서 RSK 네트워크 선택하십시오.",
@@ -180,5 +182,13 @@
   "reverse_resolution": "역순 결정",
   "set_reverse": "역순 설정",
   "has_reverse": "이미 설정됨",
-  "notifications_new_reverse_resolution": "신규 역순 결정 등록 완료"
+  "notifications_new_reverse_resolution": "신규 역순 결정 등록 완료",
+  "notifications_migrated_domain": "여러분의 도메인이 새로운 fifs registrar로 이동되었습니다",
+  "domain_not_available": "사용할 수 없습니다",
+  "start_registration_for": "다음에 대한 등록 절차 시작하기",
+  "rental_period": "대여 기간(햇수)",
+  "commit": "등록 요청하기",
+  "notifications_registrar_committed": "여러분의 도메인 등록이 요청되었습니다",
+  "notifications_registrar_revealed":  "여러분의 도메인이 등록되었습니다",
+  "login_explanation": "방금 등록된 도메인으로 이제 로그인할 수 있습니다"
 }
