+++ src/languages/pt_br.json
@@ -31,11 +31,13 @@
   "reveal_your_bid": "revelar seu lance",
   "learn_about_process": "saber mais sobre o processo",
   "hide": "ocultar",
-  "process": "Qualquer usuário pode iniciar um leilão para qualquer nome de domínio disponível. O leilão é realizado de maneira pública, respeitando os princípios de leilão de Vickrey. Um leilão de Vickrey é um tipo de leilão às cegas. Os proponentes enviam ofertas por escrito sem saber o lance dos outros participantes do leilão. Quem dá o maior lance ganha, mas o preço pago é referente ao segundo maior lance. Há 4 etapas a serem seguidas:",
-  "process_step_1": "Iniciar um leilão referente a um domínio",
-  "process_step_2": "Dar um lance no leilão",
-  "process_step_3": "Revelar o lance",
-  "process_step_4": "Encerrar o leilão",
+  "process": "Usuários podem registrar qualquer nome de domínio disponível. O registro de um nome é feito em 3 etapas",
+  "process_step_1": "Solicitar registro",
+  "process_step_1_explanation": "Sua carteira se abrirá, e lhe será solicitado que confirme a primeira de duas transações exigidas para registro.",
+  "process_step_2": "Aguardar um minuto",
+  "process_step_2_explanation": "Aguardar cerca de um minuto: o tempo de espera é exigido para verificar se outra pessoa não tentou registrar o mesmo nome e proteger você após sua solicitação.",
+  "process_step_3": "Registrar",
+  "process_step_3_explanation": "Clique em “register” [registrar], e sua carteira será reaberta. Somente após a confirmação da 2ª transação, você saberá se conseguiu o domínio",
   "learn_more": "saiba mais",
   "mycrypto_search_title": "Procure o status no MyCrypto",
   "mycrypto_select_network": "Selecione a rede da RSK no seletor localizado na parte superior direita da tela.",
@@ -180,5 +182,13 @@
   "reverse_resolution": "resolução inversa",
   "set_reverse": "definir reversão",
   "has_reverse": "já definida",
-  "notifications_new_reverse_resolution": "nova resolução inversa registrada"
+  "notifications_new_reverse_resolution": "nova resolução inversa registrada",
+  "notifications_migrated_domain": "seu domínio migrou para o novo fifs registrar",
+  "domain_not_available": "indisponível",
+  "start_registration_for": "iniciar o registro para",
+  "rental_period": "tempo de aluguel em anos",
+  "commit": "solicitação de registro",
+  "notifications_registrar_committed": "o registro de seu domínio foi solicitado",
+  "notifications_registrar_revealed":  "seu domínio foi registrado",
+  "login_explanation": "agora você pode fazer login com o domínio que acaba de registrar"
 }
