+++ src/languages/ru.json
@@ -31,11 +31,13 @@
   "reveal_your_bid": "раскрыть вашу ставку",
   "learn_about_process": "узнать о процессе",
   "hide": "скрыть",
-  "process": "Любой пользователь может начать аукцион для любого доступного доменного имени. Это открытый аукцион, в котором соблюдаются принципы аукциона Викри. Аукцион Викри — это один из видов аукциона с закрытыми предложениями. Участник подает письменную заявку, не зная о том, что предложили другие участники торгов. Побеждает самая высокая ставка, но цена определяется на основе ставки, следующей за самой высокой. Необходимо пройти 4 этапа:",
-  "process_step_1": "Начать аукцион по домену",
-  "process_step_2": "Сделать ставку на аукционе",
-  "process_step_3": "Открыть ставку",
-  "process_step_4": "Завершить аукцион",
+  "process": "Любой пользователь может зарегистрировать любое доступное доменное имя. Для регистрации имени необходимо выполнить 3 шага",
+  "process_step_1": "Отправить запрос на регистрацию",
+  "process_step_1_explanation": "Ваш кошелек откроется, и вам будет предложено подтвердить первую из двух транзакций, необходимых для регистрации.",
+  "process_step_2": "Немного подождите",
+  "process_step_2_explanation": "Подождите около минуты: период ожидания необходим для того, чтобы другой пользователь не попытался зарегистрировать то же имя, и для вашей защиты после отправки запроса.",
+  "process_step_3": "Зарегистрироваться",
+  "process_step_3_explanation": "Нажмите «Зарегистрироваться», и ваш кошелек откроется еще раз. Вы узнаете, получили ли вы домен, только после подтверждения второй транзакции",
   "learn_more": "узнать больше",
   "mycrypto_search_title": "Статус поиска на MyCrypto",
   "mycrypto_select_network": "Выберите сеть RSK в правом верхнем углу.",
@@ -180,5 +182,13 @@
   "reverse_resolution": "обратное преобразование",
   "set_reverse": "задать как обратное",
   "has_reverse": "уже используется",
-  "notifications_new_reverse_resolution": "зарегистрировано новое обратное преобразование"
+  "notifications_new_reverse_resolution": "зарегистрировано новое обратное преобразование",
+  "notifications_migrated_domain": "ваш домен был перенесен на нового регистратора fifs",
+  "domain_not_available": "недоступно",
+  "start_registration_for": "начать регистрацию для",
+  "rental_period": "срок аренды в годах",
+  "commit": "отправить запрос на регистрацию",
+  "notifications_registrar_committed": "вы запросили домен для регистрации",
+  "notifications_registrar_revealed":  "ваш домен зарегистрирован",
+  "login_explanation": "теперь вы можете войти в систему с только что зарегистрированным доменом"
 }
