+++ src/languages/zh_cn.json
@@ -31,11 +31,13 @@
   "reveal_your_bid": "开标您的出价",
   "learn_about_process": "了解流程",
   "hide": "隐藏",
-  "process": "任何用户都可以开始任何可用域名的拍卖。这是一场遵守 Vickrey 拍卖原则的公开拍卖。Vickrey 拍卖是一种密封拍卖。出价人在不知道拍卖中其他人出价的情况下提交书面出价。出价最高的人获胜，但支付的价格为第二高出价。按以下 4 个步骤执行：",
-  "process_step_1": "开始域名拍卖",
-  "process_step_2": "拍卖中的出价",
-  "process_step_3": "启封出价",
-  "process_step_4": "完成拍卖",
+  "process": "任何用户都可注册任何可用的域名。注册名称需要完成 3 个步骤",
+  "process_step_1": "要求注册",
+  "process_step_1_explanation": "将打开您的钱包，并要求您确认注册所需的两笔交易中的第一笔。",
+  "process_step_2": "等一下",
+  "process_step_2_explanation": "等待大约一分钟：需要等待一段时间，以确保他人没有尝试注册相同的名字并在您提出请求后保护您。",
+  "process_step_3": "注册",
+  "process_step_3_explanation": "点击“注册”，您的钱包将重新打开。只有在确认第二笔交易后，您才能知道是否拥有域名",
   "learn_more": "了解更多信息",
   "mycrypto_search_title": "MyCrypto 的检索状态",
   "mycrypto_select_network": "在右上角选择器上选择 RSK 网络。",
@@ -180,5 +182,13 @@
   "reverse_resolution": "反向分辨率",
   "set_reverse": "设置反向",
   "has_reverse": "已经设定",
-  "notifications_new_reverse_resolution": "注册了新的反向分辨率"
+  "notifications_new_reverse_resolution": "注册了新的反向分辨率",
+  "notifications_migrated_domain": "您的域已迁移到新的 fifs 注册商",
+  "domain_not_available": "无法使用",
+  "start_registration_for": "开始注册",
+  "rental_period": "年租期",
+  "commit": "要求注册",
+  "notifications_registrar_committed": "您的域名已被要求注册",
+  "notifications_registrar_revealed":  "您的域名已注册",
+  "login_explanation": "现在您可使用您刚注册的域登录"
 }
